fix: card selection
